feature(framework): the config is now build from both command line options and env
fixes #0
